Move links files into own folder, start regex link provider
