+++ src/vs/workbench/contrib/terminal/test/browser/links/terminalWebLinkProvider.test.ts
@@ -4,8 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as assert from 'assert';
-import { convertLinkRangeToBuffer, TerminalWebLinkProvider } from 'vs/workbench/contrib/terminal/browser/terminalWebLinkProvider';
-import { IBufferLine, IBufferCell, Terminal, ILink, IBufferRange, IBufferCellPosition } from 'xterm';
+import { TerminalWebLinkProvider } from 'vs/workbench/contrib/terminal/browser/links/terminalWebLinkProvider';
+import { Terminal, ILink, IBufferRange, IBufferCellPosition } from 'xterm';
 
 suite('Workbench - TerminalWebLinkProvider', () => {
 	suite('TerminalWebLinkProvider', () => {
@@ -85,179 +85,4 @@ suite('Workbench - TerminalWebLinkProvider', () => {
 			await assertLink('aa  https://foo.bar/[this is foo site]  aa', { range: [[5, 1], [38, 1]], text: 'https://foo.bar/[this is foo site]' });
 		});
 	});
-	suite('convertLinkRangeToBuffer', () => {
-		test('should convert ranges for ascii characters', () => {
-			const lines = createBufferLineArray([
-				{ text: 'AA http://t', width: 11 },
-				{ text: '.com/f/', width: 8 }
-			]);
-			const bufferRange = convertLinkRangeToBuffer(lines, 11, { startColumn: 4, startLineNumber: 1, endColumn: 19, endLineNumber: 1 }, 0);
-			assert.deepEqual(bufferRange, {
-				start: { x: 4, y: 1 },
-				end: { x: 7, y: 2 }
-			});
-		});
-		test('should convert ranges for wide characters before the link', () => {
-			const lines = createBufferLineArray([
-				{ text: 'A文 http://', width: 11 },
-				{ text: 't.com/f/', width: 9 }
-			]);
-			const bufferRange = convertLinkRangeToBuffer(lines, 11, { startColumn: 4, startLineNumber: 1, endColumn: 19, endLineNumber: 1 }, 0);
-			assert.deepEqual(bufferRange, {
-				start: { x: 4 + 1, y: 1 },
-				end: { x: 7 + 1, y: 2 }
-			});
-		});
-		test('should convert ranges for wide characters inside the link', () => {
-			const lines = createBufferLineArray([
-				{ text: 'AA http://t', width: 11 },
-				{ text: '.com/文/', width: 8 }
-			]);
-			const bufferRange = convertLinkRangeToBuffer(lines, 11, { startColumn: 4, startLineNumber: 1, endColumn: 19, endLineNumber: 1 }, 0);
-			assert.deepEqual(bufferRange, {
-				start: { x: 4, y: 1 },
-				end: { x: 7 + 1, y: 2 }
-			});
-		});
-		test('should convert ranges for wide characters before and inside the link', () => {
-			const lines = createBufferLineArray([
-				{ text: 'A文 http://', width: 11 },
-				{ text: 't.com/文/', width: 9 }
-			]);
-			const bufferRange = convertLinkRangeToBuffer(lines, 11, { startColumn: 4, startLineNumber: 1, endColumn: 19, endLineNumber: 1 }, 0);
-			assert.deepEqual(bufferRange, {
-				start: { x: 4 + 1, y: 1 },
-				end: { x: 7 + 2, y: 2 }
-			});
-		});
-		test('should convert ranges for ascii characters (link starts on wrapped)', () => {
-			const lines = createBufferLineArray([
-				{ text: 'AAAAAAAAAAA', width: 11 },
-				{ text: 'AA http://t', width: 11 },
-				{ text: '.com/f/', width: 8 }
-			]);
-			const bufferRange = convertLinkRangeToBuffer(lines, 11, { startColumn: 15, startLineNumber: 1, endColumn: 30, endLineNumber: 1 }, 0);
-			assert.deepEqual(bufferRange, {
-				start: { x: 4, y: 2 },
-				end: { x: 7, y: 3 }
-			});
-		});
-		test('should convert ranges for wide characters before the link (link starts on wrapped)', () => {
-			const lines = createBufferLineArray([
-				{ text: 'AAAAAAAAAAA', width: 11 },
-				{ text: 'A文 http://', width: 11 },
-				{ text: 't.com/f/', width: 9 }
-			]);
-			const bufferRange = convertLinkRangeToBuffer(lines, 11, { startColumn: 15, startLineNumber: 1, endColumn: 30, endLineNumber: 1 }, 0);
-			assert.deepEqual(bufferRange, {
-				start: { x: 4 + 1, y: 2 },
-				end: { x: 7 + 1, y: 3 }
-			});
-		});
-		test('should convert ranges for wide characters inside the link (link starts on wrapped)', () => {
-			const lines = createBufferLineArray([
-				{ text: 'AAAAAAAAAAA', width: 11 },
-				{ text: 'AA http://t', width: 11 },
-				{ text: '.com/文/', width: 8 }
-			]);
-			const bufferRange = convertLinkRangeToBuffer(lines, 11, { startColumn: 15, startLineNumber: 1, endColumn: 30, endLineNumber: 1 }, 0);
-			assert.deepEqual(bufferRange, {
-				start: { x: 4, y: 2 },
-				end: { x: 7 + 1, y: 3 }
-			});
-		});
-		test('should convert ranges for wide characters before and inside the link', () => {
-			const lines = createBufferLineArray([
-				{ text: 'AAAAAAAAAAA', width: 11 },
-				{ text: 'A文 http://', width: 11 },
-				{ text: 't.com/文/', width: 9 }
-			]);
-			const bufferRange = convertLinkRangeToBuffer(lines, 11, { startColumn: 15, startLineNumber: 1, endColumn: 30, endLineNumber: 1 }, 0);
-			assert.deepEqual(bufferRange, {
-				start: { x: 4 + 1, y: 2 },
-				end: { x: 7 + 2, y: 3 }
-			});
-		});
-		test('should convert ranges for several wide characters before the link', () => {
-			const lines = createBufferLineArray([
-				{ text: 'A文文AAAAAA', width: 11 },
-				{ text: 'AA文文 http', width: 11 },
-				{ text: '://t.com/f/', width: 11 }
-			]);
-			const bufferRange = convertLinkRangeToBuffer(lines, 11, { startColumn: 15, startLineNumber: 1, endColumn: 30, endLineNumber: 1 }, 0);
-			// This test ensures that the start offset is applies to the end before it's counted
-			assert.deepEqual(bufferRange, {
-				start: { x: 4 + 4, y: 2 },
-				end: { x: 7 + 4, y: 3 }
-			});
-		});
-		test('should convert ranges for several wide characters before and inside the link', () => {
-			const lines = createBufferLineArray([
-				{ text: 'A文文AAAAAA', width: 11 },
-				{ text: 'AA文文 http', width: 11 },
-				{ text: '://t.com/文', width: 11 },
-				{ text: '文/', width: 3 }
-			]);
-			const bufferRange = convertLinkRangeToBuffer(lines, 11, { startColumn: 15, startLineNumber: 1, endColumn: 31, endLineNumber: 1 }, 0);
-			// This test ensures that the start offset is applies to the end before it's counted
-			assert.deepEqual(bufferRange, {
-				start: { x: 4 + 4, y: 2 },
-				end: { x: 2, y: 4 }
-			});
-		});
-	});
 });
-
-const TEST_WIDE_CHAR = '文';
-const TEST_NULL_CHAR = 'C';
-
-function createBufferLineArray(lines: { text: string, width: number }[]): IBufferLine[] {
-	let result: IBufferLine[] = [];
-	lines.forEach((l, i) => {
-		result.push(new TestBufferLine(
-			l.text,
-			l.width,
-			i + 1 !== lines.length
-		));
-	});
-	return result;
-}
-
-class TestBufferLine implements IBufferLine {
-	constructor(
-		private _text: string,
-		public length: number,
-		public isWrapped: boolean
-	) {
-
-	}
-	getCell(x: number): IBufferCell | undefined {
-		// Create a fake line of cells and use that to resolve the width
-		let cells: string = '';
-		let offset = 0;
-		for (let i = 0; i <= x - offset; i++) {
-			const char = this._text.charAt(i);
-			cells += char;
-			if (this._text.charAt(i) === TEST_WIDE_CHAR) {
-				// Skip the next character as it's width is 0
-				cells += TEST_NULL_CHAR;
-				offset++;
-			}
-		}
-		return {
-			getChars: () => {
-				return x >= cells.length ? '' : cells.charAt(x);
-			},
-			getWidth: () => {
-				switch (cells.charAt(x)) {
-					case TEST_WIDE_CHAR: return 2;
-					case TEST_NULL_CHAR: return 0;
-					default: return 1;
-				}
-			}
-		} as any;
-	}
-	translateToString(): string {
-		throw new Error('Method not implemented.');
-	}
-}
